Refactor BookList with code previously on BookDetails
