+++ src/containers/BooksList.js
@@ -1,20 +1,37 @@
-import BooksDetail from './BooksDetail';
+/* eslint-disable */
+import React, { Component } from 'react';
+import { connect } from 'react-redux';
+import Book from '../components/Book';
 
-const BooksList = () => (
-  <div>
-    <table>
-      <tbody>
-        <tr>
-          <th>Id</th>
-          <th>Title</th>
-          <th>Category</th>
-        </tr>
+class BooksList extends Component {
+  render() {
+    const bookArr = Object.values(this.props.books.books);
+    return (
+      <div>
+        <table>
+          <tbody>
+            <tr>
+              <th>Id</th>
+              <th>Title</th>
+              <th>Category</th>
+            </tr>
 
-        <BooksDetail />
+            {bookArr.map(book => (
+              <Book key={book.id} id={book.id} title={book.title} category={book.category} />
+            ))}
 
-      </tbody>
-    </table>
-  </div>
-);
+          </tbody>
+        </table>
+      </div>
 
-export default BooksList;
+    );
+  }
+}
+
+function mapStateToProps(state) {
+  return ({
+    books: state.books,
+  });
+}
+
+export default connect(mapStateToProps)(BooksList);
