+++ src/reducers/index.js
@@ -1,9 +1,7 @@
 import { combineReducers } from 'redux';
-import booksData from './booksData';
 import booksReducer from './books';
 
 const rootReducer = combineReducers({
-  data: booksData,
   books: booksReducer,
 });
